Remove the "helper" class, it ended up being a bag full of everything
