+++ tests/Usages/ClassConstantUsagesTest.php
@@ -7,8 +7,9 @@ use PHPStan\File\FileHelper;
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
 use Spaze\PHPStan\Rules\Disallowed\DisallowedConstantFactory;
-use Spaze\PHPStan\Rules\Disallowed\DisallowedHelper;
 use Spaze\PHPStan\Rules\Disallowed\IsAllowedFileHelper;
+use Spaze\PHPStan\Rules\Disallowed\RuleErrors\DisallowedConstantRuleErrors;
+use Spaze\PHPStan\Rules\Disallowed\Type\TypeResolver;
 
 class ClassConstantUsagesTest extends RuleTestCase
 {
@@ -16,8 +17,9 @@ class ClassConstantUsagesTest extends RuleTestCase
 	protected function getRule(): Rule
 	{
 		return new ClassConstantUsages(
-			new DisallowedHelper(new IsAllowedFileHelper(new FileHelper(__DIR__))),
+			new DisallowedConstantRuleErrors(new IsAllowedFileHelper(new FileHelper(__DIR__))),
 			new DisallowedConstantFactory(),
+			new TypeResolver(),
 			[
 				[
 					'class' => '\Inheritance\Base',
